@@ -5,7 +5,7 @@ homepage:
 publish_to: none
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.4.0
   flutter: ">=1.17.0"
 
 dependencies:
