@@ -3,7 +3,7 @@ description: Work manager experimentations common code
 version: 0.0.1
 publish_to: none
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ^3.4.0
   flutter: ">=1.17.0"
 
 dependencies:
